requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-embed-video==1.4.9
easy-thumbnails==2.8.5
Flask==2.2.5
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
MarkupSafe==2.1.3
MouseInfo==0.1.3
paperclip==2.6.1
Pillow==9.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.0
pytweening==1.0.7
pywhatkit @ git+https://github.com/JonasSeifried/PyWhatKit@9d7fe0863ae91d358202ed7872a64813d9f819a1
requests==2.28.2
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.18
Werkzeug==3.0.1
wikipedia==1.4.0
